It now should correctly work for older versions

+++ src/main/java/io/github/spigotcvn/remapper/tasks/RemapJarTask.java
@@ -1,6 +1,7 @@
 package io.github.spigotcvn.remapper.tasks;
 
 import io.github.spigotcvn.remapper.CVNRemapper;
+import io.github.spigotcvn.remapper.util.MinecraftVersion;
 import io.github.spigotcvn.remapper.util.RemapUtil;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
@@ -12,6 +13,8 @@ import java.nio.file.Files;
 import java.nio.file.StandardCopyOption;
 
 public class RemapJarTask implements ITask {
+    public static final MinecraftVersion FIRST_MOJMAP_VERSION = new MinecraftVersion("1.17");
+
     private CVNRemapper plugin;
 
     @Override
@@ -22,7 +25,7 @@ public class RemapJarTask implements ITask {
             // it always runs after the Jar task
             project.getTasks().getByName("jar").finalizedBy(task);
             // it will run the downloadMappings task before running this task
-            task.dependsOn(plugin.getTasks().get("downloadMappings"));
+            task.dependsOn(plugin.getTasks().get("generateMappings"));
 
             task.doLast(t -> {
                 CVNRemapper.detectMinecraftVersion(project);
@@ -46,10 +49,6 @@ public class RemapJarTask implements ITask {
                         CVNRemapper.MINECRAFT_VERSION + "-R0.1-SNAPSHOT/" +
                         "minecraft-server-" + CVNRemapper.MINECRAFT_VERSION + "-R0.1-SNAPSHOT-maps-mojang.txt"
         );
-        if (!mojMaps.exists()) {
-            throw new RuntimeException("Could not find the mojang mappings file in the local maven repository\n" +
-                    "Please make sure you have ran the BuildTools with the --remapped option");
-        }
 
         File spigotMappings = new File(
                 System.getProperty("user.home") +
@@ -59,23 +58,6 @@ public class RemapJarTask implements ITask {
                         CVNRemapper.MINECRAFT_VERSION + "-R0.1-SNAPSHOT/" +
                         "minecraft-server-" + CVNRemapper.MINECRAFT_VERSION + "-R0.1-SNAPSHOT-maps-spigot.csrg"
         );
-        if (!spigotMappings.exists()) {
-            throw new RuntimeException("Could not find the spigot mappings file in the local maven repository\n" +
-                    "Please make sure you have ran the BuildTools with the --remapped option");
-        }
-
-//        File spigotMemberMappings = new File(
-//                System.getProperty("user.home") +
-//                        "/.m2/repository/" +
-//                        CVNRemapper.SPIGOT_GROUP.replace(".", "/") +
-//                        "/minecraft-server/" +
-//                        CVNRemapper.MINECRAFT_VERSION + "-R0.1-SNAPSHOT/" +
-//                        "minecraft-server-" + CVNRemapper.MINECRAFT_VERSION + "-R0.1-SNAPSHOT-maps-spigot-members.csrg"
-//        );
-//        if (!spigotMemberMappings.exists()) {
-//            throw new RuntimeException("Could not find the spigot member mappings file in the local maven repository\n" +
-//                    "Please make sure you have ran the BuildTools with the --remapped option");
-//        }
 
         File classpathJar = new File(
                 System.getProperty("user.home") +
@@ -103,34 +85,42 @@ public class RemapJarTask implements ITask {
                 StandardCopyOption.REPLACE_EXISTING
         );
 
-        File resultTmp = new File(plugin.getTmpDir(), jarTask.getArchiveFile().get().getAsFile().getName().replace(".jar", "-intermediary.jar"));
-        File officialTmp = new File(plugin.getTmpDir(), jarTask.getArchiveFile().get().getAsFile().getName().replace(".jar", "-official.jar"));
-        File spigotTmp = new File(plugin.getTmpDir(), jarTask.getArchiveFile().get().getAsFile().getName().replace(".jar", "-spigot.jar"));
 
-        plugin.setFinalTmpJar(resultTmp);
+        if(plugin.getUsesMojmaps()) {
+            File resultTmp = new File(plugin.getTmpDir(), jarTask.getArchiveFile().get().getAsFile().getName().replace(".jar", "-intermediary.jar"));
+            File officialTmp = new File(plugin.getTmpDir(), jarTask.getArchiveFile().get().getAsFile().getName().replace(".jar", "-official.jar"));
+            File spigotTmp = new File(plugin.getTmpDir(), jarTask.getArchiveFile().get().getAsFile().getName().replace(".jar", "-spigot.jar"));
 
-        plugin.addFileToMove(originalTmp, plugin.getLibsDir());
-        plugin.addFileToMove(spigotTmp, plugin.getLibsDir());
+            plugin.setFinalTmpJar(resultTmp);
 
-        RemapUtil.remapJarToObfuscated(mojMaps, jarTask.getArchiveFile().get().getAsFile(), officialTmp);
+            plugin.addFileToMove(originalTmp, plugin.getLibsDir());
+            plugin.addFileToMove(spigotTmp, plugin.getLibsDir());
 
-        RemapUtil.remapJarToSpigot(
-                plugin.getMappingsDir(),
-                spigotMappings,
-//                spigotMemberMappings,
-                mojMaps,
-                new File(plugin.getMappingsDir(), "mappings-" + CVNRemapper.MINECRAFT_VERSION + "-spigot-fields.csrg"),
-                officialTmp,
-                spigotTmp
-        );
+            RemapUtil.remapJarToObfuscated(mojMaps, jarTask.getArchiveFile().get().getAsFile(), officialTmp);
+            RemapUtil.remapJarToSpigotClass(
+                    spigotMappings,
+                    officialTmp,
+                    spigotTmp
+            );
+            RemapUtil.remapJarToIntermediary(classpathJar.toPath(), spigotTmp, resultTmp);
 
-        RemapUtil.remapJarToIntermediary(classpathJar.toPath(), officialTmp, resultTmp);
+            System.out.println("Finished remapping jars. Jar:");
+            System.out.println("Intermediary mapped (to be used with CVN): " + resultTmp.getName());
+            System.out.println("Spigot mapped (to be used like normal): " + spigotTmp.getName());
+            System.out.println("Original (Unmapped): " + originalTmp.getName());
 
-        System.out.println("Finished remapping jars. Jar:");
-        System.out.println("Intermediary mapped (to be used with CVN): " + resultTmp.getName());
-        System.out.println("Spigot mapped (to be used like normal): " + spigotTmp.getName());
-        System.out.println("Original (Unmapped): " + originalTmp.getName());
+            Files.deleteIfExists(officialTmp.toPath());
+        } else {
+            File resultTmp = new File(plugin.getTmpDir(), jarTask.getArchiveFile().get().getAsFile().getName().replace(".jar", "-intermediary.jar"));
 
-        Files.deleteIfExists(officialTmp.toPath());
+            plugin.setFinalTmpJar(resultTmp);
+            plugin.addFileToMove(originalTmp, plugin.getLibsDir());
+
+            RemapUtil.remapJarToIntermediary(classpathJar.toPath(), originalTmp, resultTmp);
+
+            System.out.println("Finished remapping jars. Jar:");
+            System.out.println("Intermediary mapped (to be used with CVN): " + resultTmp.getName());
+            System.out.println("Original (Spigot mapped): " + originalTmp.getName());
+        }
     }
 }
